Implement a better method for user selecting directories on macOS.
Got rid of the jenky mechanism used before for File Query Widget, and added something proper to asyncronously start a file picker from the C++, and then call back into the C++ when done.
This was prompted by being able to add a path to the reference spectrum widget.

+++ src/RefSpectraModel.cpp
@@ -151,11 +151,11 @@ Wt::WModelIndex RefSpectraModel::addBaseDirectory( const std::string &path )
     populateNode( node.get() );
     
     // Notify that we've added a new root node
-    beginInsertRows( Wt::WModelIndex(), m_rootNodes.size(), m_rootNodes.size() );
+    beginInsertRows( Wt::WModelIndex(), static_cast<int>(m_rootNodes.size()), m_rootNodes.size() );
     m_rootNodes.push_back( std::move(node) );
     endInsertRows();
 
-    return createIndex( m_rootNodes.size() - 1, 0, m_rootNodes.back().get() );
+    return createIndex( static_cast<int>(m_rootNodes.size() - 1), 0, m_rootNodes.back().get() );
   } catch( const std::exception &e ) {
     std::cerr << "Error adding base directory " << path << ": " << e.what() << std::endl;
   }
